move iconstate stuff into its own file
