+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod dirs;
 pub mod error;
 pub mod icon;
+pub mod iconstate;
